mail_mcp/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mail-mcp"
version = "0.1.0"
description = "Mail MCP server implementation"
requires-python = ">=3.10"
dependencies = [
  "hydra-core>=1.3,<2.0",
  "mcp>=1.9.4,<2.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4,<9.0",
]

[project.scripts]
mail-mcp = "mail_mcp.main:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"mail_mcp.conf" = ["*.yaml"]
